requirements.txt
apiclient==1.0.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
datefinder==0.7.0
DateTime==4.3
ez-setup==0.9
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.16.0
idna==2.8
JPype1==0.7.1
nltk==3.4.5
oauth2client==4.1.3
oauthlib==3.1.0
pprint==0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
python-dateutil==2.8.1
python-telegram-bot==12.3.0
pytz==2019.3
redis==3.3.11
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
setuptools-scm==3.4.1
six==1.14.0
sutime==0.2.0
telegram==0.0.1
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.7
zope.interface==4.7.1